sidenav breakout of the code
